// frontend/src/api/models/ViewMatchDetailsV2.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ViewMatchMMRCalculationDetails } from './ViewMatchMMRCalculationDetails';
import {
    ViewMatchMMRCalculationDetailsFromJSON,
    ViewMatchMMRCalculationDetailsFromJSONTyped,
    ViewMatchMMRCalculationDetailsToJSON,
} from './ViewMatchMMRCalculationDetails';
import type { ViewMatchTeamV2 } from './ViewMatchTeamV2';
import {
    ViewMatchTeamV2FromJSON,
    ViewMatchTeamV2FromJSONTyped,
    ViewMatchTeamV2ToJSON,
} from './ViewMatchTeamV2';

/**
 * 
 * @export
 * @interface ViewMatchDetailsV2
 */
export interface ViewMatchDetailsV2 {
    /**
     * 
     * @type {string}
     * @memberof ViewMatchDetailsV2
     */
    date: string;
    /**
     * 
     * @type {ViewMatchMMRCalculationDetails}
     * @memberof ViewMatchDetailsV2
     */
    mmrCalculations?: ViewMatchMMRCalculationDetails;
    /**
     * 
     * @type {ViewMatchTeamV2}
     * @memberof ViewMatchDetailsV2
     */
    team1: ViewMatchTeamV2;
    /**
     * 
     * @type {ViewMatchTeamV2}
     * @memberof ViewMatchDetailsV2
     */
    team2: ViewMatchTeamV2;
}

/**
 * Check if a given object implements the ViewMatchDetailsV2 interface.
 */
export function instanceOfViewMatchDetailsV2(value: object): boolean {
    if (!('date' in value)) return false;
    if (!('team1' in value)) return false;
    if (!('team2' in value)) return false;
    return true;
}

export function ViewMatchDetailsV2FromJSON(json: any): ViewMatchDetailsV2 {
    return ViewMatchDetailsV2FromJSONTyped(json, false);
}

export function ViewMatchDetailsV2FromJSONTyped(json: any, ignoreDiscriminator: boolean): ViewMatchDetailsV2 {
    if (json == null) {
        return json;
    }
    return {
        
        'date': json['date'],
        'mmrCalculations': json['mmrCalculations'] == null ? undefined : ViewMatchMMRCalculationDetailsFromJSON(json['mmrCalculations']),
        'team1': ViewMatchTeamV2FromJSON(json['team1']),
        'team2': ViewMatchTeamV2FromJSON(json['team2']),
    };
}

export function ViewMatchDetailsV2ToJSON(value?: ViewMatchDetailsV2 | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'date': value['date'],
        'mmrCalculations': ViewMatchMMRCalculationDetailsToJSON(value['mmrCalculations']),
        'team1': ViewMatchTeamV2ToJSON(value['team1']),
        'team2': ViewMatchTeamV2ToJSON(value['team2']),
    };
}